+++ app/pb_kits/playbook/pb_list/docs/_list_dark_lg.jsx
@@ -2,18 +2,21 @@ import React from 'react'
 
 import { List, ListItem } from '../../'
 
-const ListDark = () => {
+const ListDarkLg = () => {
   return (
     <>
       <List
           dark
+          size="large"
       >
         <ListItem>{'Item 1'}</ListItem>
         <ListItem>{'Item 2'}</ListItem>
         <ListItem>{'Item 3'}</ListItem>
+        <ListItem>{'Item 4'}</ListItem>
+        <ListItem>{'Item 5'}</ListItem>
       </List>
     </>
   )
 }
 
-export default ListDark
+export default ListDarkLg