+++ app/pb_kits/playbook/pb_title/title.rb
@@ -13,7 +13,7 @@ module Playbook
                   values: [1, 2, 3, 4],
                   default: 3
       prop :tag, type: Playbook::Props::Enum,
-                 values: %w[h1 h2 h3 h4 h5 h6 p div span],
+                 values: %w[h1 h2 h3 h4 h5 h6 div],
                  default: "h3"
       prop :text