+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renovate config updates ([#872](https://github.com/powerhome/playbook/pull/872) @thestephenmarshall)
 
 ## [4.18.1] 2020-6-15
+- Add toggling for legend click on bar and line graph
 
 ### Updated
 - Increase Playbook Compilation Performance ([#866](https://github.com/powerhome/playbook/pull/866) @thestephenmarshall)