+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+- Added rails/react full page example for filter and table ([#862](https://github.com/powerhome/playbook/pull/862) @evalouderback)
+
 ## Unreleased
 
 ### Updated